cambio de tipo  numerico para las comisiones

+++ formato-evaluacion/database/migrations/2024_06_07_154956_create_users_responses_form2_2.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreign('email')->references('email')->on('users')->onDelete('cascade');
             $table->decimal('hours', 8, 2);
             // Change the data type of puntajeEvaluar to decimal with precision 8 and scale 2
-            $table->decimal('actv2Comision', 8, 2);
+            $table->integer('actv2Comision');
 
             $table->string('obs2');
             $table->string('obs2_2');
